Renan

+++ src/main/java/com/betha/statustce/statustce/model/SituacaoEntidade.java
@@ -23,17 +23,17 @@ public class SituacaoEntidade {
 
     @Id
     @Column(name="id")
-    private Integer id;
+    private Long id;
 
-    public SituacaoEntidade(Integer id) {
+    public SituacaoEntidade(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
