+++ src/main/java/com/betha/statustce/statustce/model/Status.java
@@ -11,20 +11,20 @@ public class Status {
 
     @Id
     @Column(name="id")
-    private Integer id;
+    private Long id;
     @Column(name="descricao")
     private String descricao;
 
-    public Status(Integer id, String descricao) {
+    public Status(Long id, String descricao) {
         this.id = id;
         this.descricao = descricao;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
